requirements.txt
flask==1.1.4
markupsafe==2.0.1
werkzeug==1.0.1
pillow==8.4.0